Fix React import issue in react-hook-utils
When using `import React from "react"` there was a problem in code that
used this r3bl-ts-utils module as a dependency. This issue occurred when
the useForceUpdateFn custom hook was used.

Here are more details on the error:
TS1259: Module '".../r3bl-ts-utils/node_modules/@types/react/index"' can
only be default-imported using the 'esModuleInterop' flag.

The fix is to use `import * as React from "react"`.