Fix useKeyboard and useNodeKeypress
Fix # 1

While implementing multi-select-input ui component, a bug in
useNodeKeypress was exposed. The problem lies in how hooks work. When
connecting an external async event emitter (like process.stdin) to a
React function component via hooks is tricky. This is because hooks are
not supposed to be called by code that is not a function component
(which Node.js's process.stdin is not). The bug arose from the listener
attached to process.stdin making calls directly into a React component
via the callback function that is passed. Essentially, this listener ->
React-component-callback-function (eg: enterPressed or uparrowPressed)
caused "memory loss" in the React component, since it could not figure
out what the state is!

Using Redux would have masked this problem, since anyone can dispatch an
action to the store (async code, middleware, etc).

The fix is to use useState in order to "bind" the non-react world of
Node.js & the react world of a hook. The Node.js listener no longer
directly calls the react-component-callback-function, rather it updates
state, and an effect is registered that responds to this state change
which actually invokes the react-component-callback-function.

Fix # 2

While digging around in the code trying to fix the bug above, another
bug got exposed. The keypress listeners attached to Node.js
process.stdin were not being detached! This is now fixed.